fix: incorrect package on generated IntegrationtTests
