+++ samples/java-eventsourced-counter/src/it/java/com/example/CounterIntegrationTest.java
@@ -2,7 +2,7 @@
  * As long as this file exists it will not be re-generated.
  * You are free to make changes to this file.
  */
-package com.example.domain;
+package com.example;
 
 import com.akkaserverless.javasdk.testkit.junit.AkkaServerlessTestKitResource;
 import com.example.CounterApi;
