+++ samples/java-eventsourced-shopping-cart/src/it/java/com/example/shoppingcart/ShoppingCartIntegrationTest.java
@@ -1,39 +1,46 @@
-/* This code was generated by Akka Serverless tooling.
- * As long as this file exists it will not be re-generated.
- * You are free to make changes to this file.
+/*
+ * Copyright 2021 Lightbend Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
-package com.example.shoppingcart.domain;
+package com.example.shoppingcart.api;
 
 import com.akkaserverless.javasdk.testkit.junit.AkkaServerlessTestKitResource;
 import com.example.shoppingcart.Main;
 import com.example.shoppingcart.ShoppingCartApi;
 import com.example.shoppingcart.ShoppingCartService;
-import io.grpc.StatusRuntimeException;
 import org.junit.ClassRule;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.ExpectedException;
 
 import java.util.List;
-import java.util.concurrent.ExecutionException;
 
 import static org.junit.Assert.assertEquals;
-import static org.hamcrest.CoreMatchers.instanceOf;
 
 // Example of an integration test calling our service via the Akka Serverless proxy
 // Run all test classes ending with "IntegrationTest" using `mvn verify -Pit`
 public class ShoppingCartIntegrationTest {
 
-  /** The test kit starts both the service container and the Akka Serverless proxy. */
+  /**
+   * The test kit starts both the service container and the Akka Serverless proxy.
+   */
   @ClassRule
   public static final AkkaServerlessTestKitResource testKit =
       new AkkaServerlessTestKitResource(Main.createAkkaServerless());
 
-  /** Use the generated gRPC client to call the service through the Akka Serverless proxy. */
   private final ShoppingCartService client;
 
   public ShoppingCartIntegrationTest() {
-    client = testKit.getGrpcClient(ShoppingCartService.class);
+    this.client = testKit.getGrpcClient(ShoppingCartService.class);
   }
 
   ShoppingCartApi.Cart getCart(String cartId) throws Exception {
@@ -56,21 +63,13 @@ public class ShoppingCartIntegrationTest {
         .get();
   }
 
-  void removeItem(String cartId, String productId, String name) throws Exception {
+  void removeItem(String cartId, String productId) throws Exception {
     client
         .removeItem(
             ShoppingCartApi.RemoveLineItem.newBuilder()
                 .setCartId(cartId)
                 .setProductId(productId)
-                .setName(name)
                 .build())
-        .toCompletableFuture()
-        .get();
-  }
-
-  void removeCart(String cartId) throws Exception {
-    client
-        .removeCart(ShoppingCartApi.RemoveShoppingCart.newBuilder().setCartId(cartId).build())
         .toCompletableFuture()
         .get();
   }
@@ -98,7 +97,8 @@ public class ShoppingCartIntegrationTest {
     assertEquals(
         "shopping cart should have expected items",
         cart.getItemsList(),
-        List.of(item("a", "Apple", 1), item("b", "Banana", 2), item("c", "Cantaloupe", 3)));
+        List.of(item("a", "Apple", 1), item("b", "Banana", 2), item("c", "Cantaloupe", 3))
+    );
   }
 
   @Test
@@ -110,32 +110,15 @@ public class ShoppingCartIntegrationTest {
     assertEquals(
         "shopping cart should have expected items",
         cart1.getItemsList(),
-        List.of(item("a", "Apple", 1), item("b", "Banana", 2)));
-    removeItem("cart3", "a", "Apple");
+        List.of(item("a", "Apple", 1), item("b", "Banana", 2))
+    );
+    removeItem("cart3", "a");
     ShoppingCartApi.Cart cart2 = getCart("cart3");
     assertEquals("shopping cart should have 1 item", 1, cart2.getItemsCount());
     assertEquals(
         "shopping cart should have expected items",
         cart2.getItemsList(),
-        List.of(item("b", "Banana", 2)));
-  }
-
-  @Rule
-  public ExpectedException exceptionRule = ExpectedException.none();
-
-  @Test
-  public void removeCart() throws Exception {
-    addItem("cart4", "a", "Apple", 42);
-    ShoppingCartApi.Cart cart1 = getCart("cart4");
-    assertEquals("shopping cart should have 1 item", 1, cart1.getItemsCount());
-    assertEquals(
-        "shopping cart should have expected items",
-        cart1.getItemsList(),
-        List.of(item("a", "Apple", 42)));
-    removeCart("cart4");
-    exceptionRule.expect(ExecutionException.class);
-    exceptionRule.expectMessage("UNKNOWN: Entity deleted");
-    exceptionRule.expectCause(instanceOf(StatusRuntimeException.class));
-    getCart("cart4");
+        List.of(item("b", "Banana", 2))
+    );
   }
 }
