+++ samples/java-replicatedentity-shopping-cart/src/it/java/com/example/shoppingcart/api/ShoppingCartIntegrationTest.java
@@ -1,46 +1,35 @@
-/*
- * Copyright 2021 Lightbend Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+/* This code was generated by Akka Serverless tooling.
+ * As long as this file exists it will not be re-generated.
+ * You are free to make changes to this file.
  */
-package com.example.shoppingcart.domain;
+package com.example.shoppingcart.api;
 
-import com.example.shoppingcart.Main;
 import com.akkaserverless.javasdk.testkit.junit.AkkaServerlessTestKitResource;
+import com.example.shoppingcart.Main;
 import com.example.shoppingcart.ShoppingCartApi;
 import com.example.shoppingcart.ShoppingCartService;
+import io.grpc.StatusRuntimeException;
 import org.junit.ClassRule;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.ExpectedException;
 
 import java.util.List;
+import java.util.concurrent.ExecutionException;
 
 import static org.junit.Assert.assertEquals;
-import static java.util.concurrent.TimeUnit.*;
+import static org.hamcrest.CoreMatchers.instanceOf;
 
 // Example of an integration test calling our service via the Akka Serverless proxy
-// Run all test classes ending with "IntegrationTest" using `mvn verify -Pfailsafe`
+// Run all test classes ending with "IntegrationTest" using `mvn verify -Pit`
 public class ShoppingCartIntegrationTest {
 
-  /**
-   * The test kit starts both the service container and the Akka Serverless proxy.
-   */
+  /** The test kit starts both the service container and the Akka Serverless proxy. */
   @ClassRule
   public static final AkkaServerlessTestKitResource testKit =
-          new AkkaServerlessTestKitResource(Main.createAkkaServerless());
+      new AkkaServerlessTestKitResource(Main.createAkkaServerless());
 
-  /**
-   * Use the generated gRPC client to call the service through the Akka Serverless proxy.
-   */
+  /** Use the generated gRPC client to call the service through the Akka Serverless proxy. */
   private final ShoppingCartService client;
 
   public ShoppingCartIntegrationTest() {
@@ -67,12 +56,13 @@ public class ShoppingCartIntegrationTest {
         .get();
   }
 
-  void removeItem(String cartId, String productId) throws Exception {
+  void removeItem(String cartId, String productId, String name) throws Exception {
     client
         .removeItem(
             ShoppingCartApi.RemoveLineItem.newBuilder()
                 .setCartId(cartId)
                 .setProductId(productId)
+                .setName(name)
                 .build())
         .toCompletableFuture()
         .get();
@@ -92,7 +82,6 @@ public class ShoppingCartIntegrationTest {
         .setQuantity(quantity)
         .build();
   }
-
 
   @Test
   public void emptyCartByDefault() throws Exception {
@@ -122,7 +111,7 @@ public class ShoppingCartIntegrationTest {
         "shopping cart should have expected items",
         cart1.getItemsList(),
         List.of(item("a", "Apple", 1), item("b", "Banana", 2)));
-    removeItem("cart3", "a");
+    removeItem("cart3", "a", "Apple");
     ShoppingCartApi.Cart cart2 = getCart("cart3");
     assertEquals("shopping cart should have 1 item", 1, cart2.getItemsCount());
     assertEquals(
@@ -130,6 +119,9 @@ public class ShoppingCartIntegrationTest {
         cart2.getItemsList(),
         List.of(item("b", "Banana", 2)));
   }
+
+  @Rule
+  public ExpectedException exceptionRule = ExpectedException.none();
 
   @Test
   public void removeCart() throws Exception {
@@ -141,6 +133,9 @@ public class ShoppingCartIntegrationTest {
         cart1.getItemsList(),
         List.of(item("a", "Apple", 42)));
     removeCart("cart4");
-    assertEquals("shopping cart should be empty", 0, getCart("cart4").getItemsCount());
+    exceptionRule.expect(ExecutionException.class);
+    exceptionRule.expectMessage("UNKNOWN: Entity deleted");
+    exceptionRule.expectCause(instanceOf(StatusRuntimeException.class));
+    getCart("cart4");
   }
 }
